+++ src/datafield.cpp
@@ -7,7 +7,7 @@
 /* Data field class which stores values against their title whilst also
  * holding relevant unit, limits and raw data scaling information
  */
-DataField::DataField(std::string n_title, double n_gain, double n_offset, int n_start, int n_bytes, int n_low, int n_upp, std::string n_unit){
+DataField::DataField(std::string n_title, double n_gain, double n_offset, int n_start, int n_bytes, double n_low, double n_upp, std::string n_unit){
 	title  = n_title;
 	gain   = n_gain;
 	offset = n_offset;
